Finalized consolidated report formatting

+++ TQM/TQM/Model/YCTestSummaryModel.cs
@@ -42,7 +42,7 @@ namespace TQM.Model
 
         public decimal standardHank { get; set; }
 
-        public int deviationPercent { get; set; }
+        public decimal deviationPercent { get; set; }
 
         public string testRemark { get; set; }
